Added a public API for plugins support
This patch adds an API for plugins support to libiio. It simply dlopens each .so
matching a certain pattern (defaulting to libiio-*.so), in a certain directory
(defaulting to /usr/lib/libiio-plugins/).
The plugin can then do things in it's module constructor, for example,
registering support for a new backend with a context factory.

Change-Id: I9e60d490f95f0d8ffbb2b9c4ec05ec2bb384cf06