[DEV] Fix elements length/bits when they include a repeat count
From kernel official patch:

The current scan element type uses the following format:
  [be|le]:[s|u]bits/storagebits[>>shift].

To specify multiple elements in this type, added a repeat value.
So new format is:
  [be|le]:[s|u]bits/storagebits{X[repeat]}[>>shift].

Here X is specifying how may times, real/storage bits are repeating.
When X is value is 0 or 1, then repeat value is not used in the format,
and it will be same as existing format.

This patch only fixes the element length and bits. It will have to
be completed for iio_channel_convert() to work correctly in case
of endianess conversion.

Change-Id: I69f9c691c2421618ddcdbfadc60b58690d7cb152
Reviewed-on: https://gerrit.parrot.biz/49583
Reviewed-by: Romain Roffé <romain.roffe@parrot.com>
Reviewed-by: Rémi Lefevre <remi.lefevre@parrot.com>